+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,6 +1,6 @@
-import { authOPtions } from '@/lib/authOptions';
+import { authOptions } from '@/lib/authOptions';
 import NextAuth from 'next-auth/next';
 
-const handler = NextAuth(authOPtions);
+const handler = NextAuth(authOptions);
 
 export { handler as GET, handler as POST };
